fix: add doc file and remove untested code. issue: https://github.com/Telenav/osrm-backend/issues/143
