+++ integration/oasis/stationfinder/stations_iterator.go
@@ -3,16 +3,3 @@ package stationfinder
 type nearbyStationsIterator interface {
 	iterateNearbyStations() <-chan ChargeStationInfo
 }
-
-// ChargeStationInfo defines charge station information
-type ChargeStationInfo struct {
-	ID       string
-	Location StationCoordinate
-	err      error
-}
-
-// StationCoordinate represents location information
-type StationCoordinate struct {
-	Lat float64
-	Lon float64
-}
